Use local timezone when calculating rise transit set
When calculating rise/transit/set we need to have a base date.
riseTime, transitTime and setTime are JulianDays from a time range
starting with the base date and ending with base date + 24h.

Before this update base date was JulianDay.midnight, which gave the
wrong pair of rise/set for midnights of timezones different from GMT.